// src/model/taxonomies/camperTypes.ts
/** 
* This file has been auto-generated by '@kontent-ai/model-generator@8.0.0'.
* 
* (c) Kontent.ai
*  
* -------------------------------------------------------------------------------
* 
* Project: 🏕️KoA Personalization
* Environment: Production
* Id: f189fcee-05bc-0069-1f34-a1d35d692446
* 
* -------------------------------------------------------------------------------
**/

/**
 * All taxonomy codename values for Camper Types
 *
 * Codename: camper_types
 * Id: 66f7ee27-f81b-4c06-aacc-e41e8d7c3ff1
 */
export const camperTypesValues = [
  "experienced_camper",
  "first_time_camper",
  "rv_owner",
  "tent_camper",
  "family_camper",
  "solo_adventurer",
] as const;

/**
 * Type representing Camper Types taxonomy
 *
 * Codename: camper_types
 * Id: 66f7ee27-f81b-4c06-aacc-e41e8d7c3ff1
 */
export type CamperTypes = (typeof camperTypesValues)[number];

/**
 * Type guard for Camper Types
 *
 * Codename: camper_types
 * Id: 66f7ee27-f81b-4c06-aacc-e41e8d7c3ff1
 */
export function isCamperTypes(value: string | undefined | null): value is CamperTypes {
  return typeof value === "string" && (camperTypesValues as readonly string[]).includes(value);
}